refactor(imports): migrate find_repo_root from clients to utils
4 callers still imported find_repo_root through the backward-compat
wrapper in clients/git_client.py. Import directly from utils.git_helpers
to complete the migration started in 7ea04734.
